first kadai

+++ app/Http/Controllers/TasksController.php
@@ -15,14 +15,21 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //メッセージ一覧を取得
-        $tasks =Task::all();
-        
-        //メッセージ一覧ビューでそれを表示
-        return view('tasks.index',[
-            'tasks' =>$tasks,
-            ]);
-        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('welcome', $data);
     }
 
     /**
@@ -32,7 +39,7 @@ class TasksController extends Controller
      */
     public function create()
     {
-        //getでtasks/createにアクセスさrた場合の「新規登録画面表示処理」
+        //getでtasks/createにアクセスれた場合の「新規登録画面表示処理」
         $task = new Task;
         
         //タスク表示ビューを表示
@@ -48,20 +55,22 @@ class TasksController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
-    {
-        //バリデーション
-        $request ->validate([
-            'status' => 'required|max:10',
+    {   
+        $data=[];
+        
+        // バリデーション
+        $request->validate([
             'content' => 'required|max:255',
+            'status' => 'required|max:10',
         ]);
-        
-        //タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-        
-        //トップページへリダイレクトさせる
+
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 
@@ -77,9 +86,9 @@ class TasksController extends Controller
         $task =Task::findOrFail($id);
         
         //タスク詳細一覧でそれを表示
-        return view ('tasks.show',[
-            'task' => $task,
-            ]);
+        return view('tasks.show',[
+            'task' => $task
+        ]);
     }
 
     /**
@@ -134,12 +143,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        //idの値でタスクを検索して取得
-        $task = Task::findOrFail($id);
-        //メッセージを削除
-        $task->delete();
-        
-        //トップページへリダイレクト
+        // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
+
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
